Completed solution

+++ TaskListApplication.Server/DTOs/TaskDto.cs
@@ -6,6 +6,6 @@
         public string Title { get; set; }
         public bool IsComplete { get; set; }
         public List<SubTaskDto> SubTasks { get; set; }
-        public byte SubTasksCount { get; set; }
+        public int SubTasksCount { get; set; }
     }
 }
